Refactor: Structure project for PyPI packaging
This commit reorganizes the project structure to make it suitable for PyPI packaging and distribution.

Key changes include:
- Introduced a main package directory `testteller_rag_agent`.
- Moved core application modules (`main.py`, `agent.py`, `config.py`, `prompts.py`) and sub-modules (`data_ingestion`, `llm`, `utils`, `vector_store`) into the `testteller_rag_agent` package directory.
- Added `__init__.py` files to `testteller_rag_agent` and its sub-directories to define them as packages.
- Updated all internal import statements to use relative imports consistent with the new package structure.
- Modified `setup.py`:
    - Updated the `google-generative-ai` dependency to remove strict version pinning, allowing pip to resolve a compatible version.
    - Added `pydantic-settings` to `install_requires` as it was identified as a missing runtime dependency during local testing.

The `setup.py` file now correctly uses `find_packages()` to locate the `testteller_rag_agent` package and its sub-packages. The entry point `testteller-rag-agent` for the command-line interface is correctly configured.

The package has been successfully built into `sdist` and `wheel` distributions and tested via local installation. Instructions for publishing to PyPI have also been outlined.
