+++ testteller_rag_agent/data_ingestion/code_loader.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import aiofiles  # For async file reading
 from git import Repo, exc as GitExc  # Renamed to avoid clash with built-in exc
 # Assuming settings.code_extensions and settings.temp_clone_dir_base are here
-from config import settings
+from ..config import settings
 
 
 logger = logging.getLogger(__name__)
