+++ testteller_rag_agent/utils/helpers.py
@@ -1,6 +1,6 @@
 import logging
 import sys
-from config import settings
+from ..config import settings
 from pythonjsonlogger import jsonlogger
