+++ testteller_rag_agent/utils/retry_helpers.py
@@ -1,6 +1,6 @@
 import logging
 from tenacity import retry, stop_after_attempt, wait_exponential, retry_if_exception_type
-from config import settings
+from ..config import settings
 
 logger = logging.getLogger(__name__)
