+++ testteller_rag_agent/vector_store/chromadb_manager.py
@@ -8,8 +8,8 @@ import os  # Added
 from chromadb.utils import embedding_functions
 from chromadb.config import Settings  # Added
 from chromadb import HttpClient  # Added
-from config import settings
-from llm.gemini_client import GeminiClient
+from ..config import settings
+from ..llm.gemini_client import GeminiClient
 
 logger = logging.getLogger(__name__)
